+++ app/lib/features/slots/domain/repositories/i_slots_repository.dart
@@ -3,5 +3,7 @@ import 'package:glina/features/slots/domain/entities/slot_entity.dart';
 abstract interface class ISlotsRepository {
   Future<List<SlotEntity>> listSlots(SlotsFilter filter);
 
+  Future<List<MasterEntity>> listMasters();
+
   Future<SlotEntity> getSlot(String slotId);
 }