+++ app/lib/features/slots/presentation/manager/slots_bloc/slots_event.dart
@@ -17,3 +17,44 @@ final class RefreshSlotsEvent extends SlotsEvent {
   @override
   List<Object?> get props => [];
 }
+
+final class ApplySlotsFilterEvent extends SlotsEvent {
+  const ApplySlotsFilterEvent(this.filter);
+
+  final SlotsFilter filter;
+
+  @override
+  List<Object?> get props => [filter];
+}
+
+final class ClearSlotsFiltersEvent extends SlotsEvent {
+  const ClearSlotsFiltersEvent();
+
+  @override
+  List<Object?> get props => [];
+}
+
+final class ToggleProgramTypeFilterEvent extends SlotsEvent {
+  const ToggleProgramTypeFilterEvent(this.type);
+
+  final ProgramType type;
+
+  @override
+  List<Object?> get props => [type];
+}
+
+final class ToggleMasterFilterEvent extends SlotsEvent {
+  const ToggleMasterFilterEvent(this.masterId);
+
+  final String masterId;
+
+  @override
+  List<Object?> get props => [masterId];
+}
+
+final class ToggleOnlyAvailableFilterEvent extends SlotsEvent {
+  const ToggleOnlyAvailableFilterEvent();
+
+  @override
+  List<Object?> get props => [];
+}